requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pytest>=6.0.0
python-dotenv>=0.19.0
click>=8.0.0  # Pour l'interface en ligne de commande
tqdm>=4.62.0  # Pour les barres de progression
reedsolo>=1.5.0  # Pour la correction d'erreurs Reed-Solomon 